Add ability to disable automatic placeholders
All of the built-in renderers include a field label, so to also automatically include a placeholder on the field is a bit of duplication. This commit adds a config option to allow disabling the automatic placeholders. They're enabled by default for backward compatibility. When automatic placeholders are disabled, individual placeholders can still be added to fields.